services/analytics-service/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Database
clickhouse-driver==0.2.6
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
